Crate frunk_derives

Frunk Derives

This library holds logic for the nice custom derives in Frunk.

Links:

Source on Github

Crates.io page

Derive Macros

Generic

Derives a Generic instance based on HList for a given Struct or Tuple Struct

LabelledGeneric

Derives a Generic instance based on Field + HList for a given Struct (Tuple Structs not supported because they have no labels)