Crate frunk_derives

Frunk Derives

This library holds logic for the nice custom derives in Frunk.

Links: 1. Source on Github 2. Crates.io page

Functions

generic
Derives a Generic instance based on HList for a given Struct or Tuple Struct

labelled_generic
Derives a Generic instance based on Field + HList for a given Struct (Tuple Structs not supported because they have no labels)